[Posts] Extract file extension methods
