+++ app/models/post.rb
@@ -65,7 +65,7 @@ class Post < ApplicationRecord
 
   IMAGE_TYPES = %i[original large preview crop]
 
-  module FileMethods
+  module PostFileMethods
     extend ActiveSupport::Concern
 
     module ClassMethods
@@ -193,34 +193,6 @@ class Post < ApplicationRecord
       else
         "fit-window"
       end
-    end
-
-    def is_image?
-      file_ext =~ /jpg|jpeg|gif|png/i
-    end
-
-    def is_png?
-      file_ext =~ /png/i
-    end
-
-    def is_gif?
-      file_ext =~ /gif/i
-    end
-
-    def is_flash?
-      file_ext =~ /swf/i
-    end
-
-    def is_webm?
-      file_ext =~ /webm/i
-    end
-
-    def is_mp4?
-      file_ext =~ /mp4/i
-    end
-
-    def is_video?
-      is_webm? || is_mp4?
     end
 
     def has_preview?
@@ -1769,6 +1741,7 @@ class Post < ApplicationRecord
     end
   end
 
+  include PostFileMethods
   include FileMethods
   include ImageMethods
   include ApprovalMethods
